backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
twilio==7.0.0
PyJWT==2.8.0 